Structure of my classes changed
Structure of my new classes changed according to the structure existing in the project
(Progress -> ProgressWrapper -> ProgressWrapperSection) changed to (Progress -> ProgressSection -> JList<ProgressSection>)
