--- jdi-light-bootstrap-tests/src/main/java/io/github/com/sections/ProgressWrapperSection.java
@@ -1,18 +1,0 @@
-package io.github.com.sections;
-
-import com.epam.jdi.light.elements.composite.Section;
-import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.ProgressBarWrapper;
-
-import java.util.List;
-
-public class ProgressWrapperSection extends Section {
-
-    @UI(".progress")
-    private List<ProgressBarWrapper> progressBarWrappersList;
-
-    public List<ProgressBarWrapper> getProgressBarWrappersList() {
-        return progressBarWrappersList;
-    }
-
-}
