+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/elements/common/Progress.java
@@ -4,7 +4,7 @@ import com.epam.jdi.light.common.JDIAction;
 import com.epam.jdi.light.elements.base.UIBaseElement;
 import com.epam.jdi.light.ui.bootstrap.asserts.ProgressAssert;
 
-public class Progress extends UIBaseElement<ProgressAssert> {
+public class Progress extends UIBaseElement<ProgressAssert>  {
     @JDIAction(value = "Get '{name}' background color")
     public String getColor() {
         return uiElement.core().css("background-color");
